restructured and updated packaging
