+++ pyproject.toml
@@ -1,0 +1,4 @@
+[build-system]
+requires = ["setuptools>=45", "wheel"]
+build-backend = "setuptools.build_meta"
+
