+++ setup.cfg
@@ -1,0 +1,17 @@
+[metadata]
+name = submitit_tools
+version = 0.1.0
+description = A package to manage submitit jobs
+author = Weird Lab
+
+
+[options]
+packages = find:
+python_requires = >=3.6
+install_requires =
+    submitit
+    tqdm
+    wandb
+
+[options.extras_require]
+test = pytest
