+++ setup.py
@@ -1,16 +1,17 @@
 from setuptools import setup, find_packages
 
-requirements = [
-    "submitit",
-    "tqdm",
-    "wandb"
-]
+setup()
+# requirements = [
+#     "submitit",
+#     "tqdm",
+#     "wandb"
+# ]
 
-setup(
-    name='submitit_tools',
-    packages=find_packages(),
-    version='0.1.0',
-    description='A package to manage submitit jobs',
-    author='Weird Lab',
-    install_requires=requirements
-)
+# setup(
+#     name='submitit_tools',
+#     packages=find_packages(),
+#     version='0.1.0',
+#     description='A package to manage submitit jobs',
+#     author='Weird Lab',
+#     install_requires=requirements
+# )
